refacto: KATA-001 improving modules visibility
